Preserve the ParentOrElse behavior when changing sampling probability via configuration.
