+++ sdk/tracing/src/test/java/io/opentelemetry/sdk/trace/config/TraceConfigTest.java
@@ -19,8 +19,8 @@ package io.opentelemetry.sdk.trace.config;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
+import io.opentelemetry.sdk.trace.Sampler;
 import io.opentelemetry.sdk.trace.Samplers;
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
 
 class TraceConfigTest {
@@ -95,13 +95,15 @@ class TraceConfigTest {
   @Test
   void updateTraceConfig_OffSamplerProbability() {
     TraceConfig traceConfig = TraceConfig.getDefault().toBuilder().setSamplerProbability(0).build();
-    assertThat(traceConfig.getSampler()).isEqualTo(Samplers.alwaysOff());
+    assertThat(traceConfig.getSampler()).isSameAs(Samplers.alwaysOff());
   }
 
   @Test
   void updateTraceConfig_OnSamplerProbability() {
     TraceConfig traceConfig = TraceConfig.getDefault().toBuilder().setSamplerProbability(1).build();
-    assertThat(traceConfig.getSampler()).isEqualTo(Samplers.alwaysOn());
+
+    Sampler sampler = traceConfig.getSampler();
+    assertThat(sampler).isEqualTo(Samplers.parentOrElse(Samplers.alwaysOn()));
   }
 
   @Test
@@ -122,88 +124,5 @@ class TraceConfigTest {
     assertThat(traceConfig.getMaxNumberOfLinks()).isEqualTo(11);
     assertThat(traceConfig.getMaxNumberOfAttributesPerEvent()).isEqualTo(1);
     assertThat(traceConfig.getMaxNumberOfAttributesPerLink()).isEqualTo(2);
-  }
-
-  public static class SystemPropertiesTest {
-
-    @AfterEach
-    public void tearDown() {
-      System.clearProperty("otel.config.sampler.probability");
-      System.clearProperty("otel.config.max.attrs");
-      System.clearProperty("otel.config.max.events");
-      System.clearProperty("otel.config.max.links");
-      System.clearProperty("otel.config.max.event.attrs");
-      System.clearProperty("otel.config.max.link.attrs");
-    }
-
-    @Test
-    void updateTraceConfig_SystemProperties() {
-      System.setProperty("otel.config.sampler.probability", "0.3");
-      System.setProperty("otel.config.max.attrs", "5");
-      System.setProperty("otel.config.max.events", "6");
-      System.setProperty("otel.config.max.links", "9");
-      System.setProperty("otel.config.max.event.attrs", "7");
-      System.setProperty("otel.config.max.link.attrs", "11");
-      TraceConfig traceConfig =
-          TraceConfig.getDefault()
-              .toBuilder()
-              .readEnvironmentVariables()
-              .readSystemProperties()
-              .build();
-      assertThat(traceConfig.getSampler()).isEqualTo(Samplers.probability(0.3));
-      assertThat(traceConfig.getMaxNumberOfAttributes()).isEqualTo(5);
-      assertThat(traceConfig.getMaxNumberOfEvents()).isEqualTo(6);
-      assertThat(traceConfig.getMaxNumberOfLinks()).isEqualTo(9);
-      assertThat(traceConfig.getMaxNumberOfAttributesPerEvent()).isEqualTo(7);
-      assertThat(traceConfig.getMaxNumberOfAttributesPerLink()).isEqualTo(11);
-    }
-
-    @Test
-    void updateTraceConfig_InvalidSamplerProbability() {
-      System.setProperty("otel.config.sampler.probability", "-1");
-      assertThrows(
-          IllegalArgumentException.class,
-          () -> TraceConfig.getDefault().toBuilder().readSystemProperties().build());
-    }
-
-    @Test
-    void updateTraceConfig_NonPositiveMaxNumberOfAttributes() {
-      System.setProperty("otel.config.max.attrs", "-5");
-      assertThrows(
-          IllegalArgumentException.class,
-          () -> TraceConfig.getDefault().toBuilder().readSystemProperties().build());
-    }
-
-    @Test
-    void updateTraceConfig_NonPositiveMaxNumberOfEvents() {
-      System.setProperty("otel.config.max.events", "-6");
-      assertThrows(
-          IllegalArgumentException.class,
-          () -> TraceConfig.getDefault().toBuilder().readSystemProperties().build());
-    }
-
-    @Test
-    void updateTraceConfig_NonPositiveMaxNumberOfLinks() {
-      System.setProperty("otel.config.max.links", "-9");
-      assertThrows(
-          IllegalArgumentException.class,
-          () -> TraceConfig.getDefault().toBuilder().readSystemProperties().build());
-    }
-
-    @Test
-    void updateTraceConfig_NonPositiveMaxNumberOfAttributesPerEvent() {
-      System.setProperty("otel.config.max.event.attrs", "-7");
-      assertThrows(
-          IllegalArgumentException.class,
-          () -> TraceConfig.getDefault().toBuilder().readSystemProperties().build());
-    }
-
-    @Test
-    void updateTraceConfig_NonPositiveMaxNumberOfAttributesPerLink() {
-      System.setProperty("otel.config.max.link.attrs", "-10");
-      assertThrows(
-          IllegalArgumentException.class,
-          () -> TraceConfig.getDefault().toBuilder().readSystemProperties().build());
-    }
   }
 }
